Add playback speed options to gtk2
Add several playback speed options to the gtk2 build. The chosen speeds
were based off of those present in the windows build. Because the
current playback speed was controlled by a sleep timer that would sleep
by a 17, 17, 16 tick cycle in order to evently get 3 frames over a 50 ms
period, we also need to add simimlar cycles for the other playback
speeds. To avoid floating-point division during the emulation loop, we
store a map of these delays to achieve a similar affect.